chore(mini-rx-store): add coverage npm task, refactor name of spec helper file
