+++ libs/mini-rx-store/src/lib/spec/undo.extension.spec.ts
@@ -6,9 +6,11 @@ import {
     CounterState,
     counterStringInitialState,
     counterStringReducer,
-    CounterStringState, createUniqueCounterReducerWithAction, resetStoreConfig,
+    CounterStringState,
+    createUniqueCounterReducerWithAction,
+    resetStoreConfig,
     store,
-} from './_spec-helpers';
+} from './_spec-helpers.spec';
 import { undo, UndoExtension } from '../extensions/undo.extension';
 import { FeatureStore } from '../feature-store';
 import { Observable } from 'rxjs';
@@ -24,9 +26,9 @@ class MyFeatureStore extends FeatureStore<CounterStringState> {
     }
 
     count(payload: string): Action {
-        return this.lastAction = this.setState((state) => ({
+        return (this.lastAction = this.setState((state) => ({
             counter: state.counter + payload,
-        }));
+        })));
     }
 
     resetCount() {
@@ -34,13 +36,13 @@ class MyFeatureStore extends FeatureStore<CounterStringState> {
     }
 
     undoLastAction() {
-      if (this.lastAction) {
-        this.undo(this.lastAction);
-      }
+        if (this.lastAction) {
+            this.undo(this.lastAction);
+        }
     }
 
     undoActions(actions: Action[]) {
-        actions.forEach(item => this.undo(item))
+        actions.forEach((item) => this.undo(item));
     }
 }
 
@@ -73,7 +75,11 @@ describe('Undo Extension', () => {
 
             featureStore.resetCount();
             const undoNum2 = featureStore.count('2');
-            const actionsToUndo: Action[] = [featureStore.count('3'), featureStore.count('4'), featureStore.count('5')];
+            const actionsToUndo: Action[] = [
+                featureStore.count('3'),
+                featureStore.count('4'),
+                featureStore.count('5'),
+            ];
             featureStore.count('6');
 
             expect(counterSpy).toHaveBeenLastCalledWith('123456');
@@ -95,12 +101,13 @@ describe('Undo Extension', () => {
 
             const spy = jest.fn();
 
-            const getCount = createSelector(createFeatureSelector<CounterState>(featureKey), state => state?.counter);
+            const getCount = createSelector(
+                createFeatureSelector<CounterState>(featureKey),
+                (state) => state?.counter
+            );
             let lastAction: Action;
 
-            store
-                .select(getCount)
-                .subscribe(spy);
+            store.select(getCount).subscribe(spy);
 
             expect(spy).toHaveBeenCalledWith(1);
 
@@ -108,7 +115,7 @@ describe('Undo Extension', () => {
             store.dispatch(action);
             store.dispatch(action);
             store.dispatch(action);
-            lastAction = {...action};
+            lastAction = { ...action };
             store.dispatch(lastAction);
 
             expect(spy).toHaveBeenCalledWith(6);
@@ -123,7 +130,7 @@ describe('Undo Extension', () => {
             StoreCore.addFeature<CounterState>(featureKey, reducer);
 
             store.dispatch(action);
-            lastAction = {...action};
+            lastAction = { ...action };
             store.dispatch(lastAction);
             expect(spy).toHaveBeenCalledWith(3);
             spy.mockReset();
