+++ libs/mini-rx-store/src/lib/spec/utils.spec.ts
@@ -1,7 +1,7 @@
 import { actions$ } from '../store';
 import { ofType } from '../utils';
 import { Action } from '../models';
-import { store } from './_spec-helpers';
+import { store } from './_spec-helpers.spec';
 
 const action1: Action = {
     type: 'updateUser',
